Refactor code to improve image caching and block selection
Previously, the code was using an `imageCache` object to cache images for different block types. This has been refactored to improve caching efficiency. The code now checks if a block type exists in the `imageCache` object before caching the corresponding image. Additionally, the code now generates a random block type and adds it to the inventory multiple times.

In the `inventory.js` file, a new method `setActive` has been added to set the currently selected block type. The method sets the `selectedIdx` and `selected` properties of the inventory object based on the given index.

In the `main.js` file, the `preLoad` function now creates a new `imageCache` object to cache images for different block types. The `imageCache` object is populated by iterating over the `config.blockTypesMap`.

The `createGUI` function in `main.js` now calls the `createCache` function to generate the image and sprite caches.

The `setup` function in `main.js` now sets an interval to continuously center the player in the middle of the screen.

The `draw` function in `main.js` has been updated to display the currently selected block type by scaling and showing an image at the current mouse position. The `getCoordinates` method is also called on the `grid` object to log the coordinates of the grid.