Fix handling of failed map loading in setup function
Previously, if there was an error loading the map in the setup function, it would cause the game to crash. This commit fixes this issue by adding a try-catch block to handle errors gracefully and prevent the game from crashing. This will allow the github pages (serverless) version to work better.